Updated for 5.0
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@23760 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/library/time/Clib/Makefile.m
@@ -1,6 +1,6 @@
 AR= lib
 CC = cl
-CFLAGS = -Ox -nologo -I$(EIFFEL4)\bench\spec\$(PLATFORM)\include
+CFLAGS = -Ox -nologo -I$(EIFFEL5)\bench\spec\$(PLATFORM)\include
 LIBS = 
 MAKE = nmake
 MV = copy
@@ -15,7 +15,7 @@ all:: clean datetime.lib
 OBJECTS = datetime.obj
 
 datetime.lib: $(OBJECTS)
-	$(RM) $@
+	if exist $@ $(RM) $@
 	$(AR) -OUT:$@ $(OBJECTS)
 	if not exist ..\spec mkdir ..\spec
 	if not exist ..\spec\msc mkdir ..\spec\msc
